+++ bench/object-spread.js
@@ -51,17 +51,61 @@ for (let i = 0; i < 1000; i++) {
   }
 }
 
+const smallObject = {
+  a: {
+    c: 1,
+  },
+  b: 2,
+}
+const anotherSmallObject = {
+  a: {
+    d: 3,
+  },
+  d: 1,
+}
+
 const tableHeader = createTableHeader(['name', 'ops/sec', 'samples'])
+const bigObjectTotalKeys = Object.keys(bigObject).length
+const smallObjectTotalKeys = Object.keys(smallObject).length
 
 suite
-  .add(`Object.assign(bigObject, anotherBigObject) - Total keys: ${Object.keys(bigObject).length} - creating new object`, function () {
-    Object.assign({}, bigObject, anotherBigObject)
+  .add(`{...bigObject} - Total keys: ${bigObjectTotalKeys}`, function() {
+    const nextObject = {...bigObject}
   })
-  .add('Object.assign(bigObject, anotherBigObject) - mutating bigObject', function () {
-    Object.assign(bigObject, anotherBigObject)
+  .add(`{...smallObject} - Total keys: ${smallObjectTotalKeys}`, function() {
+    const nextObject = {...smallObject}
   })
+  .add(
+    `Object.assign({}, bigObject, anotherBigObject) - Total keys: ${
+      bigObjectTotalKeys
+    } - creating new object`,
+    function () {
+      Object.assign({}, bigObject, anotherBigObject)
+    }
+  )
+  .add(
+    'Object.assign(bigObject, anotherBigObject) - mutating bigObject',
+    function () {
+      Object.assign(bigObject, anotherBigObject)
+    }
+  )
   .add('{ ...bigObject, ...anotherBigObject }', function () {
     const nextObject = { ...bigObject, ...anotherBigObject }
+  })
+  .add(
+    'Object.assign({}, smallObject, anotherSmallObject) - creating new object',
+    function () {
+      Object.assign({}, smallObject, anotherSmallObject)
+    }
+  )
+  .add(
+    'Object.assign(smallObject, anotherSmallObject) - mutating smallObject',
+    function () {
+      Object.assign(smallObject, anotherSmallObject)
+    }
+  )
+  .add('{ ...smallObject, ...anotherSmallObject }', function () {
+    const nextObject = { ...smallObject, ...anotherSmallObject }
   })
   .on('cycle', function (event) {
     console.log(eventToMdTable(event))
